+++ src/main/java/com/api/ecoshieldwebservice/interfaces/IFeedbackServices.java
@@ -1,15 +1,15 @@
 package com.api.ecoshieldwebservice.interfaces;
 
-import com.api.ecoshieldwebservice.dtos.FeedbackRequestDTO;
+import com.api.ecoshieldwebservice.dtos.request.FeedbackRequestDTO;
 import com.api.ecoshieldwebservice.dtos.FeedbackResponseDTO;
 import com.api.ecoshieldwebservice.entities.Usuario;
 
 import java.util.List;
 
 public interface IFeedbackServices {
-    public FeedbackResponseDTO findById(Integer id);
+    public FeedbackResponseDTO findById(Long id);
     public FeedbackResponseDTO registrar(FeedbackRequestDTO dto);
-    public void borrar(Integer id);
+    public void borrar(Long id);
     public List<FeedbackResponseDTO> findAll();
     public List<FeedbackResponseDTO> findByUsuarioid(Usuario usuarioId);
     public List<FeedbackResponseDTO> findByFeedbacktipo(String tipo);
